+++ Apptentive/Apptentive/Message Center/Controllers/ApptentiveMessageCenterViewController.m
@@ -97,7 +97,6 @@ typedef NS_ENUM(NSInteger, ATMessageCenterState) {
 @property (readonly, nonatomic) NSDictionary *bodyLengthDictionary;
 
 @property (assign, nonatomic) CGRect lastKnownKeyboardRect;
-@property (assign, nonatomic) BOOL iOSAfter8_0;
 
 @end
 
@@ -126,10 +125,12 @@ typedef NS_ENUM(NSInteger, ATMessageCenterState) {
 
 	[self updateSendButtonEnabledStatus];
 
-	self.iOSAfter8_0 = [[NSProcessInfo processInfo] isOperatingSystemAtLeastVersion:(NSOperatingSystemVersion){8, 1, 0}];
-
 	self.tableView.estimatedRowHeight = 66.0;
 	self.tableView.rowHeight = UITableViewAutomaticDimension;
+
+	ApptentiveProgressNavigationBar *navigationBar = (ApptentiveProgressNavigationBar *)self.navigationController.navigationBar;
+
+	navigationBar.progressView.hidden = YES;
 }
 
 - (void)dealloc {
@@ -436,9 +437,7 @@ typedef NS_ENUM(NSInteger, ATMessageCenterState) {
 
 - (void)tableView:(UITableView *)tableView willDisplayHeaderView:(UIView *)view forSection:(NSInteger)section {
 	UITableViewHeaderFooterView *headerView = (UITableViewHeaderFooterView *)view;
-	if (self.iOSAfter8_0) {
-		headerView.textLabel.font = [self.viewModel.styleSheet fontForStyle:ApptentiveTextStyleMessageDate];
-	}
+	headerView.textLabel.font = [self.viewModel.styleSheet fontForStyle:ApptentiveTextStyleMessageDate];
 	headerView.textLabel.textColor = [self.viewModel.styleSheet colorForStyle:ApptentiveTextStyleMessageDate];
 }
 
@@ -449,6 +448,7 @@ typedef NS_ENUM(NSInteger, ATMessageCenterState) {
 			// Fall through
 		case ATMessageCenterMessageTypeMessage:
 			cell.contentView.backgroundColor = [self.viewModel.styleSheet colorForStyle:ApptentiveColorMessageBackground];
+			cell.backgroundColor = [self.viewModel.styleSheet colorForStyle:ApptentiveColorMessageBackground];
 			break;
 
 		case ATMessageCenterMessageTypeCompoundReply:
@@ -456,10 +456,12 @@ typedef NS_ENUM(NSInteger, ATMessageCenterState) {
 			// Fall through
 		case ATMessageCenterMessageTypeReply:
 			cell.contentView.backgroundColor = [self.viewModel.styleSheet colorForStyle:ApptentiveColorReplyBackground];
+			cell.backgroundColor = [self.viewModel.styleSheet colorForStyle:ApptentiveColorReplyBackground];
 			break;
 
 		case ATMessageCenterMessageTypeContextMessage:
 			cell.contentView.backgroundColor = [self.viewModel.styleSheet colorForStyle:ApptentiveColorContextBackground];
+			cell.backgroundColor = [self.viewModel.styleSheet colorForStyle:ApptentiveColorContextBackground];
 	}
 }