+++ Apptentive/Apptentive/Persistence/ApptentiveBackend.h
@@ -20,6 +20,13 @@ extern NSString *const ApptentiveAuthenticationDidFailNotificationKeyErrorType;
 extern NSString *const ApptentiveAuthenticationDidFailNotificationKeyErrorMessage;
 extern NSString *const ApptentiveAuthenticationDidFailNotificationKeyConversationIdentifier;
 
+typedef NS_ENUM(NSInteger, ApptentiveBackendState) {
+	ApptentiveBackendStateStarting,
+	ApptentiveBackendStateWaitingForDataProtectionUnlock,
+	ApptentiveBackendStatePayloadDatabaseAvailable,
+	ApptentiveBackendStateShuttingDown
+};
+
 @class ApptentiveConversation, ApptentiveEngagementManifest, ApptentiveAppConfiguration, ApptentiveMessageCenterViewController, ApptentiveMessageManager, ApptentivePayloadSender;
 
 /**
@@ -49,8 +56,6 @@ extern NSString *const ApptentiveAuthenticationDidFailNotificationKeyConversatio
 @property (readonly, strong, nonatomic) NSString *supportDirectoryPath;
 @property (strong, nullable, nonatomic) UIViewController *presentedMessageCenterViewController;
 
-@property (readonly, nonatomic) NSURLCache *imageCache;
-
 @property (copy, nullable, nonatomic) NSDictionary *currentCustomData;
 @property (copy, nonatomic) ApptentiveAuthenticationFailureCallback authenticationFailureCallback;
 
@@ -71,6 +76,7 @@ extern NSString *const ApptentiveAuthenticationDidFailNotificationKeyConversatio
 @property (readonly, strong, nonatomic) NSString *apptentiveSignature;
 @property (readonly, strong, nonatomic) NSURL *baseURL;
 @property (readonly, strong, nonatomic) NSString *storagePath;
+@property (readonly, assign, nonatomic) ApptentiveBackendState state;
 
 /**
  Instructs the serial network queue to add network operations for the currently-queued network payloads.
@@ -97,8 +103,6 @@ extern NSString *const ApptentiveAuthenticationDidFailNotificationKeyConversatio
 
 - (void)messageCenterEnteredForeground;
 - (void)messageCenterLeftForeground;
-
-- (BOOL)isReady;
 
 - (void)schedulePersonUpdate;
 - (void)scheduleDeviceUpdate;